@@ -6,29 +6,29 @@
     },
     "private": true,
     "dependencies": {
-        "@angular/common": "^13.0.0",
-        "@angular/compiler": "^13.0.0",
-        "@angular/core": "^13.0.0",
-        "@angular/forms": "^13.0.0",
-        "@angular/platform-browser": "^13.0.0",
-        "@angular/platform-browser-dynamic": "^13.0.0",
+        "@angular/common": "^14.3.0",
+        "@angular/compiler": "^14.3.0",
+        "@angular/core": "^14.3.0",
+        "@angular/forms": "^14.3.0",
+        "@angular/platform-browser": "^14.3.0",
+        "@angular/platform-browser-dynamic": "^14.3.0",
         "rxjs": "~7.4.0",
         "tslib": "^2.3.0",
         "zone.js": "~0.11.4"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "^13.3.3",
-        "@angular/cli": "^13.0.1",
-        "@angular/compiler-cli": "^13.0.0",
+        "@angular-devkit/build-angular": "^14.2.13",
+        "@angular/cli": "^14.2.13",
+        "@angular/compiler-cli": "^14.3.0",
         "@types/jasmine": "~3.10.0",
-        "@types/node": "^12.11.1",
+        "@types/node": "^22.0.0",
         "jasmine-core": "~3.10.0",
         "karma": "~6.3.0",
         "karma-chrome-launcher": "~3.1.0",
         "karma-coverage": "~2.0.3",
         "karma-jasmine": "~4.0.0",
         "karma-jasmine-html-reporter": "~1.7.0",
-        "ng-packagr": "^13.0.0",
-        "typescript": "~4.4.3"
+        "ng-packagr": "^14.2.2",
+        "typescript": "~4.8.4"
     }
 }
