@@ -30,9 +30,9 @@
         "postversion": "git commit -am \"New version\" && git push origin --tags"
     },
     "peerDependencies": {
-        "@angular/common": "^13.0.0",
-        "@angular/core": "^13.0.0",
-        "@angular/forms": "^13.0.0",
+        "@angular/common": "^14.0.0",
+        "@angular/core": "^14.0.0",
+        "@angular/forms": "^14.0.0",
         "moment": ">=2.28.0"
     },
     "dependencies": {
