@@ -1,4 +1,4 @@
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 
 import { base64 } from './validator';
 
@@ -6,12 +6,12 @@ describe('Base64', () => {
     const error = { base64: true };
 
     it('"ZGFua29nYWk=" should be base64', () => {
-        const control = new FormControl('ZGFua29nYWk=');
+        const control = new UntypedFormControl('ZGFua29nYWk=');
         expect(base64(control)).toBeNull();
     });
 
     it('"ZGFua" should not be base64', () => {
-        const control = new FormControl('ZGFua');
+        const control = new UntypedFormControl('ZGFua');
         expect(base64(control)).toEqual(error);
     });
 });
