@@ -1,4 +1,4 @@
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 
 import { digits } from './validator';
 
@@ -6,22 +6,22 @@ describe('Digits', () => {
     const error = { digits: true };
 
     it('"234" should be digits', () => {
-        const control = new FormControl('234');
+        const control = new UntypedFormControl('234');
         expect(digits(control)).toBeNull();
     });
 
     it('234 should be digits', () => {
-        const control = new FormControl(234);
+        const control = new UntypedFormControl(234);
         expect(digits(control)).toBeNull();
     });
 
     it('"abc" should not be digits', () => {
-        const control = new FormControl('abc');
+        const control = new UntypedFormControl('abc');
         expect(digits(control)).toEqual(error);
     });
 
     it('"123a" should not be digits', () => {
-        const control = new FormControl('123a');
+        const control = new UntypedFormControl('123a');
         expect(digits(control)).toEqual(error);
     });
 });
