@@ -1,4 +1,4 @@
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 
 import { email } from './validator';
 
@@ -6,17 +6,17 @@ describe('Email', () => {
     const error = { email: true };
 
     it('"test@gmail.com" should be email', () => {
-        const control = new FormControl('test@gmail.com');
+        const control = new UntypedFormControl('test@gmail.com');
         expect(email(control)).toBeNull();
     });
 
     it('"test@xxx" should not be email', () => {
-        const control = new FormControl('test');
+        const control = new UntypedFormControl('test');
         expect(email(control)).toEqual(error);
     });
 
     it('"abc" should not be email', () => {
-        const control = new FormControl('abc');
+        const control = new UntypedFormControl('abc');
         expect(email(control)).toEqual(error);
     });
 });
