@@ -1,4 +1,4 @@
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 
 import { json } from './validator';
 
@@ -6,12 +6,12 @@ describe('JSON', () => {
     const error = { json: true };
 
     it('"{"name": "xxx"}" should be json', () => {
-        const control = new FormControl('{"name": "xxx"}');
+        const control = new UntypedFormControl('{"name": "xxx"}');
         expect(json(control)).toBeNull();
     });
 
     it('"123a" should not be json', () => {
-        const control = new FormControl('123a');
+        const control = new UntypedFormControl('123a');
         expect(json(control)).toEqual(error);
     });
 });
