@@ -1,5 +1,5 @@
 import {
-    AbstractControl, FormControl, NgModel, ValidationErrors, ValidatorFn, Validators
+    AbstractControl, UntypedFormControl, NgModel, ValidationErrors, ValidatorFn, Validators
 } from '@angular/forms';
 
 import { isDate, isPresent, parseDate } from '../util/lang';
@@ -8,7 +8,7 @@ export const maxDate = (maxInput: any): ValidatorFn => {
     let value: any;
     let subscribe = false;
     let maxValue = maxInput;
-    const isForm = maxInput instanceof FormControl || maxInput instanceof NgModel;
+    const isForm = maxInput instanceof UntypedFormControl || maxInput instanceof NgModel;
 
     return (control: AbstractControl): ValidationErrors => {
         if (!subscribe && isForm) {
