@@ -1,5 +1,5 @@
 import {
-    AbstractControl, FormControl, NgModel, ValidationErrors, ValidatorFn, Validators
+    AbstractControl, UntypedFormControl, NgModel, ValidationErrors, ValidatorFn, Validators
 } from '@angular/forms';
 
 import { isDate, isPresent, parseDate } from '../util/lang';
@@ -8,7 +8,7 @@ export const minDate = (minInput: any): ValidatorFn => {
     let value: any;
     let subscribe = false;
     let minValue = minInput;
-    const isForm = minInput instanceof FormControl || minInput instanceof NgModel;
+    const isForm = minInput instanceof UntypedFormControl || minInput instanceof NgModel;
 
     return (control: AbstractControl): ValidationErrors => {
 
