@@ -1,9 +1,9 @@
-import { FormControl, ValidatorFn } from '@angular/forms';
+import { UntypedFormControl, ValidatorFn } from '@angular/forms';
 
 import { rangeLength } from './validator';
 
 describe('RangeLength [4,9],', () => {
-    let control: FormControl;
+    let control: UntypedFormControl;
     let validator: ValidatorFn;
 
     beforeEach(() => {
@@ -11,22 +11,22 @@ describe('RangeLength [4,9],', () => {
     });
 
     it('"abc" should equal to "{rangeLength: true}"', () => {
-        control = new FormControl('abc');
+        control = new UntypedFormControl('abc');
         expect(validator(control)).toEqual({ rangeLength: { value: [4, 9] } });
     });
 
     it('"abcd" should equal to "null"', () => {
-        control = new FormControl('abcd');
+        control = new UntypedFormControl('abcd');
         expect(validator(control)).toBeNull();
     });
 
     it('"abcdefghi" should equal to "null"', () => {
-        control = new FormControl('abcdefghi');
+        control = new UntypedFormControl('abcdefghi');
         expect(validator(control)).toBeNull();
     });
 
     it('"abcdefghij" should equal to "{rangeLength: true}"', () => {
-        control = new FormControl('abcdefghij');
+        control = new UntypedFormControl('abcdefghij');
         expect(validator(control)).toEqual({ rangeLength: { value: [4, 9] } });
     });
 });
